% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppock.R
\name{coppock_curve}
\alias{coppock_curve}
\title{Coppock curve}
\usage{
coppock_curve(x, n1 = 14L, n2 = 11L, n3 = 10L)
}
\arguments{
\item{x}{A vector}

\item{n1}{Rate of change period #1}

\item{n2}{Rate of change period #2}

\item{n3}{Weighted moving average period}
}
\value{
A vector
}
\description{
A momentum indicator, originally a "Very Long Term (VLT) Momentum Index",
defined as the 10-month weighted moving average (WMA) of the sum of a series'
11-month and 14-month rates of change (ROC).
}
\details{
Using the default parameters, the Coppock curve is defined as the
\code{10-WMA of (ROC[11] + ROC[14])}.
}
