% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppock.R
\name{coppock_curve_solve_each}
\alias{coppock_curve_solve_each}
\title{Solve Coppock curve for each \code{x}}
\usage{
coppock_curve_solve_each(x, n1 = 14L, n2 = 11L, n3 = 10L, y = 0)
}
\arguments{
\item{x}{A vector}

\item{n1}{Rate of change period #1}

\item{n2}{Rate of change period #2}

\item{n3}{Weighted moving average period}

\item{y}{Coppock curve value}
}
\value{
A vector
}
\description{
Solve for each value of \code{x} that gives Coppock curve value \code{y}.
}
\details{
The calculation of the Coppock curve is such that changing one value of
\code{x} would affect the Coppock curve at other points. Therefore, the
solution for each \code{x} is independent and is found by holding constant
all other values of \code{x}.
}
