% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coppock.R
\name{coppock_curve_solve_next}
\alias{coppock_curve_solve_next}
\title{Solve Coppock curve for next \code{x}}
\usage{
coppock_curve_solve_next(x, n1 = 14L, n2 = 11L, n3 = 10L, y = 0)
}
\arguments{
\item{x}{A vector}

\item{n1}{Rate of change period #1}

\item{n2}{Rate of change period #2}

\item{n3}{Weighted moving average period}

\item{y}{Coppock curve value}
}
\value{
A numeric
}
\description{
Solve for next value of \code{x} that would give Coppock curve value
\code{y}.
}
